make MID 8bit

+++ System/Constants.h
@@ -7,7 +7,7 @@ namespace RTE {
 	// TODO: Look into not using distinctive types for IDs.
 	typedef uint_fast32_t MOID; //!< Distinctive type definition for MovableObject IDs.
 	// TODO Use MIDs everywhere for materials
-	typedef uint_fast32_t MID; //!< Distinctive type definition for Material IDs.
+	typedef uint_fast8_t MID; //!< Distinctive type definition for Material IDs.
 #pragma endregion
 
 #pragma region Game Version
